feat(poidh): refactor data fetching to SSR with service layer
- Create src/services/poidh.ts with React cache() for server-side
  deduplication and ISR caching (15min list, 1min detail)
- Extract BountiesView client component from page.tsx with initialData
  support for seamless SSR → client hydration
- Extract BountyDetailView client component with initialData from
  server-fetched bounty data
- Convert list page to async Server Component with Suspense fallback
- Convert detail page to SSR with generateMetadata for dynamic SEO
- Thin API routes from ~100 lines to ~20 lines, delegating to service
- Add initialData support to usePoidhBounties hook
- Export BountyGridSkeleton for Suspense fallback
