+++ src/components/bounties/BountyGrid.tsx
@@ -69,3 +69,13 @@ export function BountyGrid({ bounties, isLoading, error }: BountyGridProps) {
     </div>
   );
 }
+
+export function BountyGridSkeleton() {
+  return (
+    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-5">
+      {Array.from({ length: 6 }).map((_, i) => (
+        <SkeletonCard key={i} />
+      ))}
+    </div>
+  );
+}
